+++ app/models/loader/types.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 require_relative '../../controllers/concerns/authorize_resource'
+require_relative '../../domain/secrets/cache/redis_handler'
 
 module Loader
   module Types
@@ -397,10 +398,14 @@ module Loader
     end
 
     class Delete < Deletion
+      include Secrets::RedisHandler
       def delete!
         if policy_object.record.respond_to?(:resourceid)
           resource = ::Resource[policy_object.record.resourceid]
-          resource.destroy if resource
+          if resource
+            resource.destroy
+            delete_redis_secret(resource.id) if resource.kind == 'variable'
+          end
         end
         if policy_object.record.respond_to?(:roleid)
           role = ::Role[policy_object.record.roleid]