+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
-## [1.0.25-cloud] - 2024-04-14
+## [1.0.26-cloud] - 2024-05-05
+### Changed
+- Text review for dynamic secrets
+
+## [1.0.25-cloud] - 2024-04-21
 ### Added
 - Return permissions in static secret
 - Get dynamic secret